+++ pytradfri/typing.py
